% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barmixR_fit-class.R
\name{barmixR_fit}
\alias{barmixR_fit}
\title{Class barmixR_fit}
\value{
An object of class \code{barmixR_fit}.
}
\description{
Object returned by \code{barmixRQTR()}.
}
\details{
Contains fitted Bayesian models and associated metadata for
barcode composition and population size.
}
\seealso{
\code{\link{barmixRQTR}}
}
